Adding tests regarding Authorization and Access Tests, working!

+++ backend/src/controllers/courseController.js
@@ -1,10 +1,16 @@
-// src/controllers/courseController.js
-
-const Course = require('../models/course'); // Actualizado para coincidir con el nombre del archivo
+const Course = require('../models/course');
 
 exports.createCourse = async (req, res) => {
   try {
-    const newCourse = new Course(req.body);
+    if (req.user.role !== 'admin' && req.user.role !== 'instructor') {
+      return res.status(403).json({ msg: 'Not authorized to create courses' });
+    }
+
+    const newCourse = new Course({
+      ...req.body,
+      instructor: req.user.id
+    });
+
     await newCourse.save();
     res.status(201).json(newCourse);
   } catch (error) {
@@ -35,11 +41,17 @@ exports.getCourse = async (req, res) => {
 
 exports.updateCourse = async (req, res) => {
   try {
-    const course = await Course.findByIdAndUpdate(req.params.id, req.body, { new: true });
+    const course = await Course.findById(req.params.id);
     if (!course) {
       return res.status(404).json({ message: 'Course not found' });
     }
-    res.status(200).json(course);
+
+    if (course.instructor.toString() !== req.user.id && req.user.role !== 'admin') {
+      return res.status(403).json({ message: 'Not authorized to update this course' });
+    }
+
+    const updatedCourse = await Course.findByIdAndUpdate(req.params.id, req.body, { new: true });
+    res.status(200).json(updatedCourse);
   } catch (error) {
     res.status(400).json({ error: error.message });
   }
@@ -47,10 +59,16 @@ exports.updateCourse = async (req, res) => {
 
 exports.deleteCourse = async (req, res) => {
   try {
-    const course = await Course.findByIdAndDelete(req.params.id);
+    const course = await Course.findById(req.params.id);
     if (!course) {
       return res.status(404).json({ message: 'Course not found' });
     }
+
+    if (course.instructor.toString() !== req.user.id && req.user.role !== 'admin') {
+      return res.status(403).json({ message: 'Not authorized to delete this course' });
+    }
+
+    await Course.findByIdAndDelete(req.params.id);
     res.status(200).json({ message: 'Course deleted successfully' });
   } catch (error) {
     res.status(400).json({ error: error.message });
